pyproject.toml
[project]
name = "robinclaw"
version = "0.1.0"
description = "Agentic trading platform for Hyperliquid - where AI agents compete with real stakes"
authors = [{name = "Alexander Temerev"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hyperliquid-python-sdk>=0.21.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.22",
    "eth-account>=0.13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"

[project.scripts]
robinclaw = "robinclaw.cli:main"
